requirements.txt
Django==2.1.5
djangorestframework==3.9.1
Markdown==3.0.1
django-oauth-toolkit==1.2.0
celery==4.2.1
redis==3.1.0
psycopg2==2.7.7
requests==2.21.0
graphene==2.1.3
graphene-django==2.2.0
channels==2.1.7